+++ packages/scrollable_positioned_list/lib/src/positioned_list.dart
@@ -309,13 +309,14 @@ class _PositionedListState extends State<PositionedList> {
     if (!updateScheduled) {
       updateScheduled = true;
       SchedulerBinding.instance.addPostFrameCallback((_) {
-        if (registeredElements.value == null) {
+        final elements = registeredElements.value;
+        if (elements == null) {
           updateScheduled = false;
           return;
         }
         final positions = <ItemPosition>[];
         RenderViewportBase? viewport;
-        for (var element in registeredElements.value!) {
+        for (var element in elements) {
           final RenderBox box = element.renderObject as RenderBox;
           viewport ??= RenderAbstractViewport.of(box) as RenderViewportBase?;
           var anchor = 0.0;
@@ -328,6 +329,8 @@ class _PositionedListState extends State<PositionedList> {
           }
 
           final ValueKey<int> key = element.widget.key as ValueKey<int>;
+          // Skip this element if `box` has never been laid out.
+          if (!box.hasSize) continue;
           if (widget.scrollDirection == Axis.vertical) {
             final reveal = viewport!.getOffsetToReveal(box, 0).offset;
             if (!reveal.isFinite) continue;
@@ -342,6 +345,7 @@ class _PositionedListState extends State<PositionedList> {
           } else {
             final itemOffset =
                 box.localToGlobal(Offset.zero, ancestor: viewport).dx;
+            if (!itemOffset.isFinite) continue;
             positions.add(ItemPosition(
                 index: key.value,
                 itemLeadingEdge: (widget.reverse