+++ packages/visibility_detector/lib/src/render_visibility_detector.dart
@@ -140,7 +140,14 @@ mixin RenderVisibilityDetectorBase on RenderObject {
     }
     bool isFirstUpdate = _updates.isEmpty;
     _updates[key] = () {
-      _fireCallback(layer, bounds);
+      if (bounds == null) {
+        // This can happen if set onVisibilityChanged was called with a non-null
+        // value but this render object has not been laid out. In that case,
+        // it has no size or geometry, and we should not worry about firing
+        // an update since it never has been visible.
+        return;
+      }
+      _fireCallback(layer, bounds!);
     };
     final updateInterval = VisibilityDetectorController.instance.updateInterval;
     if (updateInterval == Duration.zero) {
@@ -165,7 +172,7 @@ mixin RenderVisibilityDetectorBase on RenderObject {
   }
 
   VisibilityInfo _determineVisibility(ContainerLayer? layer, Rect bounds) {
-    if (_disposed || layer?.attached == false || !attached) {
+    if (_disposed || layer == null || layer.attached == false || !attached) {
       // layer is detached and thus invisible.
       return VisibilityInfo(
         key: key,
@@ -174,39 +181,50 @@ mixin RenderVisibilityDetectorBase on RenderObject {
     }
     final transform = Matrix4.identity();
 
-    // Create a list of RenderObjects from this to the root, excluding the root
-    // since that has the DPR transform and we want to work with logical pixels.
-    // Cannot use the layer tree since some ancestor render object may have
-    // directly transformed/clipped the canvas. If there is some way to figure
-    // out how to get the RenderObjects below [layer], could take advantage of
-    // the usually shallower height of the layer tree compared to the render
-    // tree. Alternatively, if the canvas itself exposed the current matrix/clip
-    // we could use that.
-    RenderObject? ancestor = parent as RenderObject?;
-
-    final List<RenderObject> ancestors = <RenderObject>[];
-    ancestors.add(this);
-    RenderObject child = this;
-    while (ancestor != null && ancestor.parent != null) {
-      if (!ancestor.paintsChild(child)) {
-        return VisibilityInfo(key: key, size: bounds.size);
+    // Check if any ancestors decided to skip painting this RenderObject.
+    if (parent != null) {
+      RenderObject ancestor = parent! as RenderObject;
+      RenderObject child = this;
+      while (ancestor.parent != null) {
+        if (!ancestor.paintsChild(child)) {
+          return VisibilityInfo(key: key, size: bounds.size);
+        }
+        child = ancestor;
+        ancestor = ancestor.parent! as RenderObject;
       }
-      ancestors.add(ancestor);
-      child = ancestor;
-      ancestor = ancestor.parent as RenderObject?;
     }
 
-    // Determine the transform and clip from first child of root down to
-    // this.
+    // Create a list of Layers from layer to the root, excluding the root
+    // since that has the DPR transform and we want to work with logical pixels.
+    // Add one extra leaf layer so that we can apply the transform of `layer`
+    // to the matrix.
+    ContainerLayer? ancestor = layer;
+    final List<ContainerLayer> ancestors = <ContainerLayer>[ContainerLayer()];
+    while (ancestor != null && ancestor.parent != null) {
+      ancestors.add(ancestor);
+      ancestor = ancestor.parent;
+    }
+
     Rect clip = Rect.largest;
     for (int index = ancestors.length - 1; index > 0; index -= 1) {
       final parent = ancestors[index];
       final child = ancestors[index - 1];
-      Rect? parentClip = parent.describeApproximatePaintClip(child);
+      Rect? parentClip = parent.describeClipBounds();
       if (parentClip != null) {
         clip = clip.intersect(MatrixUtils.transformRect(transform, parentClip));
       }
-      parent.applyPaintTransform(child, transform);
+      parent.applyTransform(child, transform);
+    }
+
+    // Apply whatever transform/clip was on the canvas when painting.
+    if (_lastPaintClipBounds != null) {
+      clip = clip.intersect(MatrixUtils.transformRect(
+        transform,
+        _lastPaintClipBounds!,
+      ));
+    }
+    if (_lastPaintTransform != null) {
+      transform.multiply(_lastPaintTransform!);
     }
     return VisibilityInfo.fromRects(
       key: key,
@@ -217,11 +235,21 @@ mixin RenderVisibilityDetectorBase on RenderObject {
 
   /// Used to get the bounds of the render object when it is time to update
   /// clients about visibility.
-  Rect get bounds;
+  ///
+  /// A null value means bounds are not available.
+  Rect? get bounds;
+
+  Matrix4? _lastPaintTransform;
+  Rect? _lastPaintClipBounds;
 
   @override
   void paint(PaintingContext context, Offset offset) {
     if (onVisibilityChanged != null) {
+      _lastPaintClipBounds = context.canvas.getLocalClipBounds();
+      _lastPaintTransform =
+          Matrix4.fromFloat64List(context.canvas.getTransform())
+            ..translate(offset.dx, offset.dy, 0);
+
       _compositionCallbackCanceller?.call();
       _compositionCallbackCanceller =
           context.addCompositionCallback((Layer layer) {
@@ -261,7 +289,7 @@ class RenderVisibilityDetector extends RenderProxyBox
   final Key key;
 
   @override
-  Rect get bounds => semanticBounds;
+  Rect? get bounds => hasSize ? semanticBounds : null;
 }
 
 /// The [RenderObject] corresponding to the [SliverVisibilityDetector] widget.
@@ -283,7 +311,11 @@ class RenderSliverVisibilityDetector extends RenderProxySliver
   final Key key;
 
   @override
-  Rect get bounds {
+  Rect? get bounds {
+    if (geometry == null) {
+      return null;
+    }
+
     Size widgetSize;
     Offset widgetOffset;
     switch (applyGrowthDirectionToAxisDirection(