+++ packages/visibility_detector/test/render_visibility_detector_test.dart
@@ -7,7 +7,6 @@
 import 'dart:ui';
 
 import 'package:flutter/rendering.dart';
-import 'package:flutter/widgets.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:visibility_detector/src/render_visibility_detector.dart';
 import 'package:visibility_detector/visibility_detector.dart';
@@ -29,6 +28,9 @@ void main() {
     detector.layout(BoxConstraints.tight(const Size(200, 200)));
     detector.paint(context, Offset.zero);
     detector.paint(context, Offset.zero);
+
+    context.stopRecordingIfNeeded(); // ignore: invalid_use_of_protected_member
+
     expect(layer.subtreeHasCompositionCallbacks, true);
 
     expect(detector.debugScheduleUpdateCount, 0);
@@ -74,6 +76,7 @@ void main() {
     expect(layer.subtreeHasCompositionCallbacks, true);
 
     expect(detector.debugScheduleUpdateCount, 0);
+    context.stopRecordingIfNeeded(); // ignore: invalid_use_of_protected_member
     layer.buildScene(SceneBuilder()).dispose();
 
     expect(detector.debugScheduleUpdateCount, 1);
@@ -99,6 +102,7 @@ void main() {
     expect(layer.subtreeHasCompositionCallbacks, false);
 
     expect(detector.debugScheduleUpdateCount, 0);
+    context.stopRecordingIfNeeded(); // ignore: invalid_use_of_protected_member
     layer.buildScene(SceneBuilder()).dispose();
 
     expect(detector.debugScheduleUpdateCount, 0);
@@ -125,8 +129,49 @@ void main() {
     expect(layer.subtreeHasCompositionCallbacks, false);
 
     expect(detector.debugScheduleUpdateCount, 0);
+    context.stopRecordingIfNeeded(); // ignore: invalid_use_of_protected_member
     layer.buildScene(SceneBuilder()).dispose();
 
     expect(detector.debugScheduleUpdateCount, 0);
+  });
+
+  testWidgets('RVS can schedule an update for a RO that is not laid out',
+      (WidgetTester tester) async {
+    final RenderVisibilityDetector detector = RenderVisibilityDetector(
+      key: Key('test'),
+      onVisibilityChanged: (_) {
+        fail('should not get called');
+      },
+    );
+
+    // Force an out of band update to get scheduled without laying out.
+    detector.onVisibilityChanged = (_) {
+      fail('This should also not get called');
+    };
+
+    expect(detector.debugScheduleUpdateCount, 1);
+
+    detector.dispose();
+  });
+
+  testWidgets(
+      'RVS (Sliver) can schedule an update for a RO that is not laid out',
+      (WidgetTester tester) async {
+    final RenderSliverVisibilityDetector detector =
+        RenderSliverVisibilityDetector(
+      key: Key('test'),
+      onVisibilityChanged: (_) {
+        fail('should not get called');
+      },
+    );
+
+    // Force an out of band update to get scheduled without laying out.
+    detector.onVisibilityChanged = (_) {
+      fail('This should also not get called');
+    };
+
+    expect(detector.debugScheduleUpdateCount, 1);
+
+    detector.dispose();
   });
 }